Winners 객체 삭제 winner 여부 Races로 이동
