+++ src/test/java/step4/RacesTest.java
@@ -1,12 +1,16 @@
 package step4;
 
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import step4.model.Car;
 import step4.model.Races;
 
 import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
@@ -28,5 +32,25 @@ public class RacesTest {
         Races races = mock(Races.class);
         races.start(condition);
         verify(races, times(1)).start(condition);
+    }
+
+    @Test
+    @DisplayName("레이싱이 완료 된 자동차 리스트를 받아 우승자를 정한다.")
+    void getWinnersTest() {
+        Races racing = new Races(List.of("jane", "john"));
+        racing.start(4);
+        assertThat(racing.getWinners().stream().map(Car::getName).collect(Collectors.joining()))
+                .isEqualTo("jane");
+
+    }
+
+    @Test
+    @DisplayName("레이싱이 완료 된 자동차 리스트를 받아 우승자가 2명 이상이라면 두명 다 조회된다.")
+    void getMultipleWinnersTest() {
+        Races racing = new Races(List.of("jane", "john"));
+        racing.start(4);
+        assertThat(racing.getWinners().stream().map(Car::getName).collect(Collectors.joining(",")))
+                .isEqualTo("jane,john");
+
     }
 }
